+++ fj-universe-test/src/test/java/test/org/fugerit/java/universe/utils/SectionModel.java
@@ -1,0 +1,18 @@
+package test.org.fugerit.java.universe.utils;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NonNull;
+
+@AllArgsConstructor
+public class SectionModel {
+
+	@NonNull @Getter private String name;
+	
+	@NonNull @Getter private String[] items;
+
+	public static SectionModel newInstance( String name, String... items ) {
+		return new SectionModel( name, items );
+	}
+	
+}